moved most of action from core/ to api/, where it belongs
